Split IsOneOfAssert
