+++ tests/IsOneOfAssert.php
@@ -5,47 +5,8 @@
  * @license http://www.yiiframework.com/license/
  */
 
-namespace yiiunit;
-
-use yii\helpers\VarDumper;
-
-/**
- * IsOneOfAssert asserts that the value is one of the expected values.
- */
-class IsOneOfAssert extends \PHPUnit\Framework\Constraint\Constraint
-{
-    private $allowedValues;
-
-    /**
-     * IsOneOfAssert constructor.
-     * @param array $allowedValues
-     */
-    public function __construct(array $allowedValues)
-    {
-        parent::__construct();
-        $this->allowedValues = $allowedValues;
-    }
-
-
-    /**
-     * Returns a string representation of the object.
-     *
-     * @return string
-     */
-    public function toString()
-    {
-        $allowedValues = array_map(function ($value) {
-            return VarDumper::dumpAsString($value);
-        }, $this->allowedValues);
-        $expectedAsString = implode(', ', $allowedValues);
-        return "is one of $expectedAsString";
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    protected function matches($other)
-    {
-        return in_array($other, $this->allowedValues, false);
-    }
+if (PHP_VERSION_ID >= 70100) {
+    require 'IsOneOfAssert7.php';
+} else {
+    require 'IsOneOfAssert5.php';
 }
